+++ crates/model/src/instruments/mod.rs
@@ -399,12 +399,12 @@ pub trait Instrument: 'static + Send {
             last_price.as_f64().is_finite(),
             "non-finite price passed to calculate_base_quantity",
         )?;
-        let quantity_decimal = Decimal::from_f64_retain(quantity.as_f64()).ok_or_else(|| {
+        let quantity_dec = Decimal::from_f64_retain(quantity.as_f64()).ok_or_else(|| {
             anyhow::anyhow!("non-finite quantity passed to calculate_base_quantity")
         })?;
-        let price_decimal = Decimal::from_f64_retain(last_price.as_f64())
+        let price_dec = Decimal::from_f64_retain(last_price.as_f64())
             .ok_or_else(|| anyhow::anyhow!("non-finite price passed to calculate_base_quantity"))?;
-        let value_decimal = (quantity_decimal / price_decimal).round_dp_with_strategy(
+        let value_decimal = (quantity_dec / price_dec).round_dp_with_strategy(
             self.size_precision().into(),
             RoundingStrategy::MidpointNearestEven,
         );